+++ core/src/test/scala/io/finch/syntax/TwitterMapperSyntaxSpec.scala
@@ -1,0 +1,10 @@
+package io.finch.syntax
+
+import com.twitter.util.Future
+import io.catbird.util._
+
+class TwitterMapperSyntaxSpec extends MapperSyntaxSpec {
+
+  it should behave like endpointMapper[Future]
+
+}
