+++ core/src/test/scala/io/finch/syntax/scala/ScalaMapperSyntaxSpec.scala
@@ -3,9 +3,9 @@ package io.finch.syntax.scala
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
-import io.finch.syntax.MapperSyntaxBehaviour
+import io.finch.syntax.MapperSyntaxSpec
 
-class ScalaMapperSyntaxSpec extends MapperSyntaxBehaviour {
+class ScalaMapperSyntaxSpec extends MapperSyntaxSpec {
 
   it should behave like endpointMapper[Future]
